disable dotnet.device due to instability

+++ +stdlib/+dotnet/ram_total.m
@@ -1,5 +1,5 @@
 function bytes = ram_total()
   % .NET is 2-3x faster than Java for this
   % https://learn.microsoft.com/en-us/dotnet/api/system.gcmemoryinfo.totalavailablememorybytes
-  bytes = System.GC.GetGCMemoryInfo().TotalAvailableMemoryBytes;
+  bytes = uint64(System.GC.GetGCMemoryInfo().TotalAvailableMemoryBytes);
 end
